move methods from utils to tgb_modules/tkg_utils and update imports
