+++ examples/linkproppred/tkgl-polecat/recurrencybaseline.py
@@ -28,7 +28,8 @@ sys.path.append(tgb_modules_path)
 from tgb_modules.recurrencybaseline_predictor import baseline_predict, baseline_predict_remote
 from tgb.linkproppred.evaluate import Evaluator
 from tgb.linkproppred.dataset import LinkPropPredDataset 
-from tgb.utils.utils import set_random_seed,  save_results, create_basis_dict, group_by, reformat_ts
+from tgb.utils.utils import set_random_seed,  save_results 
+from tgb_modules.tkg_utils import create_basis_dict, group_by, reformat_ts
 
 def predict(num_processes,  data_c_rel, all_data_c_rel, alpha, lmbda_psi,
             perf_list_all, hits_list_all, window, neg_sampler, split_mode):
