+++ tgb/datasets/dataset_scripts/compute_dataset_stats.py
@@ -3,23 +3,17 @@ import numpy as np
 import sys
 import os
 import os.path as osp
-from pathlib import Path
 tgb_modules_path = osp.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..'))
 sys.path.append(tgb_modules_path)
 import json
 
 ## imports
-
-
 import numpy as np
-import seasonal
-
-import pandas as pd
 from datetime import datetime
 #internal imports 
 from tgb.linkproppred.dataset import LinkPropPredDataset 
-from tgb.utils.utils import  reformat_ts, get_original_ts
-import tgb.utils.dataset_utils as du
+from tgb_modules.tkg_utils import  reformat_ts
+import tgb.datasets.dataset_scripts.dataset_utils as du
 
 import networkx as nx
 import matplotlib.pyplot as plt
